+++ packages/adp-tooling/src/index.ts
@@ -1,6 +1,11 @@
 export * from './types';
+export * from './prompts';
 export * from './preview/adp-preview';
 export { generate, migrate } from './writer';
 export { generateChange } from './writer/editors';
 export { promptGeneratorInput, PromptDefaults } from './base/prompt';
+<<<<<<< HEAD
 export { getPromptsForChangeInbound } from './prompts';
+=======
+export { getManifest } from './base/abap';
+>>>>>>> main