+++ packages/adp-tooling/src/prompts/index.ts
@@ -1,1 +1,5 @@
+<<<<<<< HEAD
 export { getPrompts as getPromptsForChangeInbound } from './change-inbound';
+=======
+export { getPrompts as getPromptsForChangeDataSource } from './change-data-source';
+>>>>>>> main