+++ packages/store/src/entities/backend-system.ts
@@ -2,13 +2,13 @@ import type { EntityKey } from '.';
 import { sensitiveData, serializable } from '../decorators';
 
 export const AuthenticationType = {
-    Basic : 'basic',
-    ReentranceTicket : 'reentranceTicket',
-    OAuth2RefreshToken : 'oauth2',
-    OAuth2ClientCredential : 'oauth2ClientCredential'
+    Basic: 'basic',
+    ReentranceTicket: 'reentranceTicket',
+    OAuth2RefreshToken: 'oauth2',
+    OAuth2ClientCredential: 'oauth2ClientCredential'
 } as const;
 
-export type AuthenticationType = typeof AuthenticationType[keyof typeof AuthenticationType];
+export type AuthenticationType = (typeof AuthenticationType)[keyof typeof AuthenticationType];
 
 export class BackendSystem {
     @serializable public readonly name: string;