+++ packages/fiori-annotation-api/CHANGELOG.md
@@ -1,5 +1,60 @@
 # @sap-ux/fiori-annotation-api
 
+## 0.1.11
+
+### Patch Changes
+
+-   Updated dependencies [c2359077]
+    -   @sap-ux/cds-annotation-parser@0.2.0
+    -   @sap-ux/cds-odata-annotation-converter@0.3.0
+    -   @sap-ux/odata-annotation-core@0.2.0
+    -   @sap-ux/odata-annotation-core-types@0.4.0
+    -   @sap-ux/odata-entity-model@0.3.0
+    -   @sap-ux/odata-vocabularies@0.4.0
+    -   @sap-ux/project-access@1.23.0
+    -   @sap-ux/xml-odata-annotation-converter@0.3.0
+
+## 0.1.10
+
+### Patch Changes
+
+-   Updated dependencies [9ea58ad4]
+    -   @sap-ux/project-access@1.22.4
+    -   @sap-ux/cds-odata-annotation-converter@0.2.9
+
+## 0.1.9
+
+### Patch Changes
+
+-   e669a2da: fix: update access modifier
+
+## 0.1.8
+
+### Patch Changes
+
+-   3afada7e: fix exception while setting annotation value to null and insert element notation for null cases in xml
+
+## 0.1.7
+
+### Patch Changes
+
+-   42b2f459: Fix compositon targets for CDS
+
+## 0.1.6
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.22.3
+-   @sap-ux/cds-odata-annotation-converter@0.2.9
+
+## 0.1.5
+
+### Patch Changes
+
+-   Updated dependencies [399d2ad8]
+    -   @sap-ux/project-access@1.22.2
+    -   @sap-ux/cds-odata-annotation-converter@0.2.9
+
 ## 0.1.4
 
 ### Patch Changes