+++ packages/odata-service-writer/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @sap-ux/odata-service-writer
 
+## 0.21.0
+
+### Minor Changes
+
+-   c2359077: [BREAKING CHANGE] Change TypeScript transpile target to ES2021 to align with NodeJS 18+
+
+### Patch Changes
+
+-   Updated dependencies [c2359077]
+    -   @sap-ux/mockserver-config-writer@0.6.0
+    -   @sap-ux/ui5-config@0.23.0
+
+## 0.20.5
+
+### Patch Changes
+
+-   Updated dependencies [1a1baeb0]
+    -   @sap-ux/ui5-config@0.22.10
+    -   @sap-ux/mockserver-config-writer@0.5.8
+
+## 0.20.4
+
+### Patch Changes
+
+-   Updated dependencies [399d2ad8]
+    -   @sap-ux/ui5-config@0.22.9
+    -   @sap-ux/mockserver-config-writer@0.5.7
+
 ## 0.20.3
 
 ### Patch Changes